Executable now has its separate CMakeLists.txt in preparation for rust stuff
