+++ srcs/gbclodo/renderer/wrapper_c/wrapper.h
